requirements.txt
# requirements.txt
langflow
docling
docling-core
docling-parse
#ransformers==4.57.1
langchain
langchain-community
langchain-groq
chromadb
faiss-cpu
tiktoken
pytesseract
pillow
python-dotenv
requests
pymupdf
fastembed
